backendのファイル構成は考え直さないといけないな

+++ run.sh
@@ -1,8 +1,10 @@
+set -e
+
 PROJECT_ROOT=$(cd "$(dirname "$0")" && pwd)
 # 仮想環境の有効化
 ACTIVATE_VENV="$PROJECT_ROOT/my-venv/bin/activate"
 # 実行中のプロセスを追跡するためのPIDファイルを各のするディレクトリ
-PID_DIR="$PROFECT_ROOT/run"
+PID_DIR="$PROJECT_ROOT/run"
 mkdir -p "$PID_DIR"
 
 PID_MQTT="$PID_DIR/mqtt_worker.pid"
@@ -15,9 +17,15 @@ cleanup(){
 
     # PIDファイルからプロセスIDを読み取り、終了させる
 
-    kill $(cat "$PID_MQTT") 2>/dev/null || true
-    kill $(cat "$PID_FASTAPI") 2>/dev/null || true
-    kill $(cat "$PID_REACT") 2>/dev/null || true
+    if [ -f "$PID_MQTT" ]; then
+        kill $(cat "$PID_MQTT") 2>/dev/null || true
+    fi
+    if [ -f "$PID_FASTAPI" ]; then
+        kill $(cat "$PID_FASTAPI") 2>/dev/null || true
+    fi
+    if [ -f "$PID_REACT" ]; then
+        kill $(cat "$PID_REACT") 2>/dev/null || true
+    fi
 
     # PIDファイルを削除
     rm -f "$PID_MQTT" "$PID_FASTAPI" "$PID_REACT"
@@ -26,7 +34,7 @@ cleanup(){
 }
 
 # crl+C が押されたcleanup関数を呼び出す
-trap cleanup SIGINT
+trap 'cleanup' SIGINT
 
 echo "---バックエンドを起動---"
 
@@ -34,7 +42,7 @@ echo "---バックエンドを起動---"
 
 if [ -f "$ACTIVATE_VENV" ]; then
     echo "仮想環境を有効化"
-    source "$ACTIVATE_VENV"
+    . "$ACTIVATE_VENV"
 
 else
     echo "仮想環境が見つかりません: $ACITVATE_VENV"
@@ -42,29 +50,31 @@ else
 fi
 
 # MQTTワーカーを起動
-echo "MQTTワーカーをバックグラウンドで起動中"
+echo "--MQTTワーカーをバックグラウンドで起動中--"
 (
-    cd "$PROJECT_ROOT/Backend"
-    python app/workers/mqtt_worker.py &
-    echo &! > "$DID_MQTT" #PIDをファイルに保存
+    cd "$PROJECT_ROOT/BackEnd"
+    python app/workers/mqtt_worker.py $ #2> /dev/null &
+    echo $! > "$PID_MQTT" #PIDをファイルに保存
 )
-
-echo "Fastapi サーバーをバックグラウンドで起動中"
+echo ""
+echo "--MQTTワーカー起動完了--"
+echo "--Fastapi サーバーをバックグラウンドで起動中--"
 (
-    cd "$PROJECT_ROOT/Backend"
+    cd "$PROJECT_ROOT/BackEnd/"
     uvicorn app.main:app --host 0.0.0.0 --port 8000 &
     echo $! > "$PID_FASTAPI" # PID をファイルに保存
 )
-
+echo ""
+echo "--fastapiサーバー起動完了--"
 echo "サーバーが起動するのを待機中"
 sleep 5 # サーバー起動を待つために数秒待機
 
-echo "フロントエンド起動"
+echo "---フロントエンド起動---"
 
-echo " react開発サーバーをバックグラウンドで起動中"
+echo "--react開発サーバーをバックグラウンドで起動中--"
 (
     cd "$PROJECT_ROOT/frontend"
-    npm run dev 2> /dev/null $
+    npm run dev $ #2> /dev/null $
     echo $! > "$PID_REACT" # pidをファイルに保存
 )
